Update global state
