+++ project_pokedex/src/components/header/Header.js
@@ -4,6 +4,7 @@ export default function Header(props){
     return(
         <Container>
             <h1>Pokédex</h1>
+            <button onClick={props.navigation}>{props.title}</button>
         </Container>
     )
 }
